# packages/contracts/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: Pal Control API
  version: 0.4.0
  license:
    name: MIT
    identifier: MIT
  description: >-
    Browser-facing API for Palworld operations. The official Palworld REST API
    and native mod bridge are internal adapters and must not be exposed directly.
    Management operations require a configured administrator API key and are
    additionally restricted to the loopback management boundary. Privileged
    economy and season writes also require a TOTP code and an auditable reason.
    Player-portal sessions use a separate cookie identity and never authorize
    management operations. Non-loopback deployments require HTTPS at a trusted
    reverse proxy in addition to these application controls.
servers:
  - url: http://127.0.0.1:5180/api/v1
    description: Local development
...
paths:
...
  /admin/player-identity-security-audit:
    get:
      tags: [AdminSecurity]
      summary: List redacted player authentication and moderation security events
      description: >-
        Owner-only reverse-chronological projection. Player subjects and source
        IP addresses are represented only by SHA-256 fingerprints. Verification
        codes, challenge IDs, session cookies, session tokens and CSRF tokens are
        never persisted in this audit.
      operationId: listPlayerIdentitySecurityAuditEvents
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Recent redacted identity-security events in reverse insertion order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerIdentitySecurityAuditList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/player-sessions/revoke:
    post:
      tags: [AdminSecurity]
      summary: Revoke all player-portal sessions for one bound player identity
      description: >-
        Owner-only revocation. The identifier may be the active world's complete
        PlayerUID or the normalized platform subject. Revocation does not ban the
        player, but every currently issued process session becomes invalid before
        the response is returned.
      operationId: revokePlayerPortalSessions
      parameters:
        - name: X-Correlation-ID
          in: header
          required: false
          description: Optional caller correlation value retained in both administrator and identity audits.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [serverId, playerIdentifier]
              properties:
                serverId:
                  type: string
                  minLength: 1
                  maxLength: 64
                playerIdentifier:
                  type: string
                  minLength: 3
                  maxLength: 128
      responses:
        "200":
          description: All matching process sessions were revoked
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [revokedSessions, resolvedSubjects, correlationId]
                properties:
                  revokedSessions:
                    type: integer
                    minimum: 0
                  resolvedSubjects:
                    type: integer
                    minimum: 1
                  correlationId:
                    type: string
                    minLength: 1
        "400":
          description: The server or player identifier was malformed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No active binding or process session matched the player identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    adminApiKey:
      type: apiKey
      in: header
      name: X-Pal-Admin-Key
      description: >-
        Administrator API key. Configure only a SHA-256 digest at rest; never
        store the plaintext key in repository configuration or logs.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    playerSessionCookie:
      type: apiKey
      in: cookie
      name: pal_player_session
      description: >-
        Opaque player session set only after successful in-game OTP verification.
        The deployment uses an HttpOnly, SameSite=Strict cookie scoped to
        /api/v1/player and enables Secure behind HTTPS. Authenticated player
        identity is derived exclusively from this cookie, never from a userId
        query parameter or business request body.
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Authenticated principal does not have the required role or security proof
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    PlayerIdentitySecurityAuditEvent:
      type: object
      additionalProperties: false
      required:
        - sequence
        - eventId
        - correlationId
        - eventType
        - outcome
        - reasonCode
        - riskLevel
        - subjectFingerprint
        - sourceIpFingerprint
        - affectedSessions
        - occurredAt
      properties:
        sequence:
          type: integer
          format: int64
          minimum: 1
        eventId:
          type: string
          format: uuid
        correlationId:
          type: string
          minLength: 1
        eventType:
          type: string
          enum:
            - login_code_request
            - login_code_verification
            - session_rejected
            - session_logout
            - administrative_session_revocation
            - administrative_ban
            - administrative_unban
        outcome:
          type: string
          enum: [succeeded, denied, failed, ignored]
        reasonCode:
          type: string
          pattern: "^[a-z0-9_-]{1,64}$"
        riskLevel:
          type: string
          enum: [normal, suspicious, high]
        subjectFingerprint:
          type: [string, "null"]
          pattern: "^[0-9A-F]{64}$"
        sourceIpFingerprint:
          type: [string, "null"]
          pattern: "^[0-9A-F]{64}$"
        affectedSessions:
          type: integer
          minimum: 0
        occurredAt:
          type: string
          format: date-time
    PlayerIdentitySecurityAuditList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlayerIdentitySecurityAuditEvent"
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: [string, "null"]
security:
  - adminApiKey: []